Add GitHub Pages static build
