+++ web/public/index.html
@@ -4,11 +4,10 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>Every-Embodied 电子书</title>
-    <link rel="icon" type="image/png" href="/raw/YXNzZXRzL-e6v-eovy5wbmc?v=1776084187000" />
-    <link rel="apple-touch-icon" href="/raw/YXNzZXRzL-e6v-eovy5wbmc?v=1776084187000" />
-    <link rel="preload" as="image" href="/raw/YXNzZXRzL-mmlumhteiDjOaZr-Wbvi5wbmc?v=1776084112000" />
-    <link rel="stylesheet" href="/vendor/katex/katex.min.css" />
-    <link rel="stylesheet" href="/styles.css" />
+    <link rel="icon" type="image/png" href="./favicon.png" />
+    <link rel="apple-touch-icon" href="./favicon.png" />
+    <link rel="stylesheet" href="./vendor/katex/katex.min.css" />
+    <link rel="stylesheet" href="./styles.css" />
   </head>
   <body class="app-booting">
     <header class="topbar">
@@ -17,9 +16,9 @@
         <span></span>
         <span></span>
       </button>
-      <a class="brand" href="#/" aria-label="回到项目首页">
+      <a class="brand" href="#/" aria-label="回到首页">
         <span class="brand-mark">
-          <img id="brandLogo" src="/raw/YXNzZXRzL-WktOWDjyBsb2dvLmpwZw" alt="Every-Embodied logo" />
+          <img id="brandLogo" src="./favicon.png" alt="Every-Embodied logo" />
         </span>
         <span>
           <strong>Every-Embodied</strong>
@@ -27,7 +26,7 @@
         </span>
       </a>
       <nav class="top-links" aria-label="快速入口">
-        <a href="#/" id="homeLink">首页</a>
+        <a href="#/">首页</a>
         <a href="#/home/module-navigation">导航模型</a>
         <a href="#/home/module-operation">操作模型</a>
         <a href="#/home/module-world">世界模型</a>
@@ -36,7 +35,14 @@
         <a href="https://github.com/datawhalechina/every-embodied" target="_blank" rel="noopener noreferrer">GitHub</a>
       </nav>
       <div class="top-meta" id="topMeta">正在读取目录</div>
-      <a class="star-pill" id="starPill" href="https://github.com/datawhalechina/every-embodied" target="_blank" rel="noopener noreferrer" aria-label="GitHub stars">
+      <a
+        class="star-pill"
+        id="starPill"
+        href="https://github.com/datawhalechina/every-embodied"
+        target="_blank"
+        rel="noopener noreferrer"
+        aria-label="GitHub stars"
+      >
         <span class="star-icon">★</span>
         <span>Starred</span>
         <strong id="starCount">--</strong>
@@ -63,11 +69,14 @@
       <main class="reader" id="reader">
         <section class="home-page" id="homePage" hidden>
           <section class="home-hero" aria-labelledby="homeTitle">
-            <img class="home-hero-image" id="homeHeroImage" alt="Every-Embodied 课程封面" />
+            <img class="home-hero-image" id="homeHeroImage" alt="Every-Embodied 首页背景" />
             <div class="home-hero-shade"></div>
             <div class="home-hero-copy">
               <p class="home-kicker">Every-Embodied</p>
-              <h1 id="homeTitle" aria-label="从 0 到 1，走进具身智能。"><span>从 0 到 1，</span><span>走进具身智能。</span></h1>
+              <h1 id="homeTitle" aria-label="从 0 到 1，走进具身智能。">
+                <span>从 0 到 1，</span>
+                <span>走进具身智能。</span>
+              </h1>
               <p>机器人基础、仿真工具、视觉感知、VLA、导航与复现路线，一本电子书连起来。</p>
               <div class="home-actions">
                 <a class="home-button primary" href="#/book">进一步了解</a>
@@ -105,7 +114,7 @@
             <h1 id="docTitle">正在加载</h1>
           </div>
           <div class="reader-actions">
-            <button class="control-button" id="toggleCatalog" type="button">章节总览</button>
+            <button class="control-button" id="toggleCatalog" type="button">展开章节总览</button>
             <button class="control-button" id="refreshCatalog" type="button">刷新目录</button>
           </div>
         </section>
@@ -114,14 +123,14 @@
           <div class="overview-copy">
             <p class="eyebrow">学习导航</p>
             <h2>按文件夹自动排布</h2>
-            <p>新增或移动 Markdown 后，刷新页面即可同步目录。</p>
+            <p>新增或移动 Markdown 后，重新构建即可同步目录。</p>
           </div>
           <div class="overview-grid" id="overviewGrid"></div>
         </section>
 
         <article class="article" id="article" aria-live="polite"></article>
 
-        <nav class="pager" aria-label="上一篇下一篇">
+        <nav class="pager" aria-label="上一篇和下一篇">
           <a class="pager-link" id="prevLink" href="#"></a>
           <a class="pager-link" id="nextLink" href="#"></a>
         </nav>
@@ -136,6 +145,6 @@
     </div>
 
     <div class="scrim" id="scrim"></div>
-    <script src="/app.js" defer></script>
+    <script src="./app.js" defer></script>
   </body>
 </html>
